fix: 수익률 계산의 결과로 소수점 둘째 자리에서 반올림 하도록 수정

+++ src/main/java/lotto/model/WinningDetails.java
@@ -4,5 +4,5 @@ import lotto.constant.LotteryRank;
 
 import java.util.Map;
 
-public record WinningDetails(Map<LotteryRank, Integer> breakdown, float returnOnInvestment) {
+public record WinningDetails(Map<LotteryRank, Integer> breakdown, double returnOnInvestment) {
 }
